# openapi/okoscope-v1.yaml
openapi: 3.1.0
info:
  title: Okoscope Web API
  version: 1.0.0
  description: Stable API contract for the separately deployed Okoscope Web UI. Timestamps are RFC 3339 UTC values.
servers:
  - url: /
security:
  - sessionAuth: []
paths:
...
  /api/v1/organization:
    get:
      operationId: getCurrentOrganization
      responses:
        {
          '200': { $ref: '#/components/responses/Organization' },
          '401': { $ref: '#/components/responses/Error' },
        }
  /api/v1/projects:
    get:
      operationId: listProjects
      parameters:
        [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }]
      responses:
        {
          '200': { $ref: '#/components/responses/ProjectPage' },
          '400': { $ref: '#/components/responses/Error' },
        }
  /api/v1/projects/{project_id}:
    parameters: [{ $ref: '#/components/parameters/ProjectId' }]
    get:
      operationId: getProject
      responses:
        {
          '200': { $ref: '#/components/responses/Project' },
          '404': { $ref: '#/components/responses/Error' },
        }
  /api/v1/projects/{project_id}/applications:
    parameters: [{ $ref: '#/components/parameters/ProjectId' }]
    get:
      operationId: listApplications
      parameters:
        [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }]
      responses:
        {
          '200': { $ref: '#/components/responses/ApplicationPage' },
          '404': { $ref: '#/components/responses/Error' },
        }
    post:
      operationId: createApplication
      security: [{ sessionAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/ProvisioningIdempotencyKey' }]
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: '#/components/schemas/CreateNamedResource' } } }
      responses:
        '201':
          {
            description: Application and one-time initial credential created,
            content:
              { application/json: { schema: { $ref: '#/components/schemas/CreatedApplication' } } },
          }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
  /api/v1/projects/{project_id}/applications/{application_id}:
    parameters:
      [
        { $ref: '#/components/parameters/ProjectId' },
        { $ref: '#/components/parameters/ApplicationId' },
      ]
    get:
      operationId: getApplication
      responses:
        {
          '200': { $ref: '#/components/responses/Application' },
          '404': { $ref: '#/components/responses/Error' },
        }
  /api/v1/projects/{project_id}/applications/{application_id}/workers:
    parameters:
      [
        { $ref: '#/components/parameters/ProjectId' },
        { $ref: '#/components/parameters/ApplicationId' },
      ]
    get:
      operationId: listApplicationWorkers
      parameters:
        [
          { $ref: '#/components/parameters/OpaqueCursor' },
          { $ref: '#/components/parameters/Limit' },
        ]
      responses:
        {
          '200': { $ref: '#/components/responses/ApplicationWorkerPage' },
          '400': { $ref: '#/components/responses/Error' },
          '401': { $ref: '#/components/responses/Error' },
          '404': { $ref: '#/components/responses/Error' },
        }
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: okoscope_session
      description: Opaque HttpOnly browser session; clients send requests with credentials enabled.
  parameters:
...
    Cursor:
      {
        name: cursor,
        in: query,
        schema: { type: string, format: uuid },
        description: Opaque cursor scoped to the authenticated collection.,
      }
    OpaqueCursor:
      {
        name: cursor,
        in: query,
        schema: { type: string },
        description: Opaque cursor scoped to the authenticated collection; clients must not parse it.,
      }
    Limit:
      { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
    ProjectId:
      { name: project_id, in: path, required: true, schema: { type: string, format: uuid } }
...
    ApplicationId:
      { name: application_id, in: path, required: true, schema: { type: string, format: uuid } }
...
    ProvisioningIdempotencyKey:
      {
        name: Idempotency-Key,
        in: header,
        required: false,
        description: Canonical UUID. Reusing it with a different request returns idempotency_key_reused. Application and credential replays never return plaintext token material.,
        schema: { type: string, format: uuid },
      }
  headers:
    RequestId: { schema: { type: string }, description: Correlation identifier for the request. }
...
  schemas:
...
    OrganizationRole: { type: string, enum: [owner, admin, member] }
    ProjectRole: { type: string, enum: [admin, member] }
    EffectiveAccessSource: { type: string, enum: [platform, organization, project] }
    AccessCapabilities:
      type: object
      additionalProperties: false
      required: [manage_platform, manage_organization, create_project, manage_project_members, create_application, manage_credentials, organization_roles_grantable, project_roles_grantable]
      properties:
        manage_platform: { type: boolean }
        manage_organization: { type: boolean }
        create_project: { type: boolean }
        manage_project_members: { type: boolean }
        create_application: { type: boolean }
        manage_credentials: { type: boolean }
        organization_roles_grantable: { type: array, uniqueItems: true, maxItems: 3, items: { $ref: '#/components/schemas/OrganizationRole' } }
        project_roles_grantable: { type: array, uniqueItems: true, maxItems: 2, items: { $ref: '#/components/schemas/ProjectRole' } }
...
    Error:
      type: object
      additionalProperties: false
      required: [error, message, request_id]
      properties:
        error:
          type: string
          enum: [validation_failed, unauthorized, invalid_credentials, email_verification_required, registration_disabled, registration_conflict, untrusted_origin, forbidden, not_found, organization_not_found, project_not_found, application_not_found, user_not_found, user_not_eligible, setup_already_completed, invalid_setup_token, setup_rate_limited, privilege_confirmation_required, current_password_invalid, self_promotion_forbidden, last_super_admin_required, last_organization_owner_required, organization_limit_reached, invitation_unusable, invitation_account_mismatch, invitation_exists, rate_limited, mail_unavailable, idempotency_key_reused, operation_already_completed, internal_error]
        message: { type: string, example: resource not found }
        request_id: { type: string, example: 0ec02ed2-8483-4981-893e-bffc535897d7 }
        fields:
          type: object
          additionalProperties: { type: string }
          description: Present for field-level validation failures.
...
    Organization:
      type: object
      additionalProperties: false
      required: [id, slug, name, created_at]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        slug: { type: string }
        name: { type: string }
        created_at: { $ref: '#/components/schemas/Timestamp' }
...
    CreateNamedResource:
      type: object
      additionalProperties: false
      required: [slug, name]
      properties:
        slug: { type: string, minLength: 1, maxLength: 63, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
        name: { type: string, minLength: 1, maxLength: 120 }
...
    ProvisionedApplication:
      type: object
      additionalProperties: false
      required: [id, organization_id, project_id, slug, name, created_at]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        organization_id: { $ref: '#/components/schemas/Uuid' }
        project_id: { $ref: '#/components/schemas/Uuid' }
        slug: { type: string }
        name: { type: string }
        created_at: { $ref: '#/components/schemas/Timestamp' }
...
    IssuedApplicationCredential:
      type: object
      additionalProperties: false
      required: [id, name, token, token_hint, created_at, shown_once]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        name: { type: string }
        token: { type: string, writeOnly: true, pattern: '^oko_app_v1_[A-Za-z0-9_-]+$' }
        token_hint: { type: string }
        created_at: { $ref: '#/components/schemas/Timestamp' }
        shown_once: { type: boolean, const: true }
    CreatedApplication:
      type: object
      additionalProperties: false
      required: [application, credential]
      properties:
        application: { $ref: '#/components/schemas/ProvisionedApplication' }
        credential: { $ref: '#/components/schemas/IssuedApplicationCredential' }
    Project:
      type: object
      additionalProperties: false
      required: [id, slug, name, created_at, archived_at, application_count, runtime_group_count, effective_project_role, effective_access_source, capabilities]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        slug: { type: string }
        name: { type: string }
        created_at: { $ref: '#/components/schemas/Timestamp' }
        archived_at: { $ref: '#/components/schemas/NullableTimestamp' }
        application_count: { type: integer, format: int64, minimum: 0 }
        runtime_group_count: { type: integer, format: int64, minimum: 0 }
        effective_project_role: { anyOf: [{ $ref: '#/components/schemas/ProjectRole' }, { type: 'null' }] }
        effective_access_source: { anyOf: [{ $ref: '#/components/schemas/EffectiveAccessSource' }, { type: 'null' }] }
        capabilities: { $ref: '#/components/schemas/AccessCapabilities' }
    ProjectPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Project' } }
        next_cursor: { $ref: '#/components/schemas/NullableUuid' }
    Application:
      type: object
      additionalProperties: false
      required:
        [
          id,
          project_id,
          slug,
          name,
          created_at,
          release_count,
          runtime_group_count,
          latest_observed_at,
          effective_project_role,
          effective_access_source,
          capabilities,
        ]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        project_id: { $ref: '#/components/schemas/Uuid' }
        slug: { type: string }
        name: { type: string }
        created_at: { $ref: '#/components/schemas/Timestamp' }
        release_count: { type: integer, format: int64, minimum: 0 }
        runtime_group_count: { type: integer, format: int64, minimum: 0 }
        latest_observed_at: { $ref: '#/components/schemas/NullableTimestamp' }
        effective_project_role: { anyOf: [{ $ref: '#/components/schemas/ProjectRole' }, { type: 'null' }] }
        effective_access_source: { anyOf: [{ $ref: '#/components/schemas/EffectiveAccessSource' }, { type: 'null' }] }
        capabilities: { $ref: '#/components/schemas/AccessCapabilities' }
    ApplicationPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Application' } }
        next_cursor: { $ref: '#/components/schemas/NullableUuid' }
...
    ApplicationWorkerPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor, coverage]
      properties:
        coverage: { $ref: '#/components/schemas/RuntimeRetentionCoverage' }
        items:
          { type: array, maxItems: 200, items: { $ref: '#/components/schemas/ApplicationWorker' } }
        next_cursor: { $ref: '#/components/schemas/NullableOpaqueCursor' }
...
    Uuid: { type: string, format: uuid }
    NullableUuid: { type: [string, 'null'], format: uuid }
    NullableOpaqueCursor: { type: [string, 'null'] }
    Timestamp: { type: string, format: date-time, example: '2026-08-17T12:00:00Z' }
    NullableTimestamp: { type: [string, 'null'], format: date-time }
...
  responses:
...
    Organization:
      {
        description: Current organization,
        headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } },
        content: { application/json: { schema: { $ref: '#/components/schemas/Organization' } } },
      }
    Project:
      {
        description: Project summary,
        headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } },
        content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } },
      }
    ProjectPage:
      {
        description: Project page,
        headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } },
        content: { application/json: { schema: { $ref: '#/components/schemas/ProjectPage' } } },
      }
    Application:
      {
        description: Application summary,
        headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } },
        content: { application/json: { schema: { $ref: '#/components/schemas/Application' } } },
      }
    ApplicationPage:
      {
        description: Application page,
        headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } },
        content: { application/json: { schema: { $ref: '#/components/schemas/ApplicationPage' } } },
      }
    ApplicationWorkerPage:
      {
        description: Application worker platform metadata page,
        headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } },
        content:
          { application/json: { schema: { $ref: '#/components/schemas/ApplicationWorkerPage' } } },
      }
...
    Error:
      description: Correlated API error
      headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } }
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }